package/apparmor: bump to version 3.0.0
This patch bumps apparmor and libapparmor to version 3.0.0

Of all our patches, one was already a backport, one is no longer
applicable (file largely rewritten), and the four others have been
applied upstream now.

The hash for the tarball is available on the homepage, so use that
as a reference.

Signed-off-by: Angelo Compagnucci <angelo@amarulasolutions.com>
[yann.morin.1998@free.fr:
  - expand commit log to explain why we drop all the patches
  - add a reference for the hash for the source tarball
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>