package/nginx: use /run instead of /var/run
This is a follow-up to 4027ba29f448 ("package/nginx: use /run for
PIDFile"), in which we missed that nginx is still built with /var/run
paths.

This commit changes the compile options to use /run instead of
/var/run for pid and lock file to make it consistent.

Further dropping the passing of the pid option in the service file as
this isn't neccessary. Neither debian nor nginx default .service does
it.

Signed-off-by: Michael Nosthoff <buildroot@heine.tech>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit d200ceffb2eab5c097684c5b9a2e6c6097b1f698)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>