package/libglib2: fix util-linux-libs dependency
Commit 2acaa86ce630f616b12aa92493e0269421c5bc97 replaced util-linux by
util-linux-libs but this raises the following build failure:

Makefile:585: *** util-linux-libs is in the dependency chain of libglib2 that has added it to its _DEPENDENCIES variable without selecting it or depending on it from Config.in.  Stop.

To fix this build failure, use util-linux or util-linux-libs depending
on BR2_PACKAGE_UTIL_LINUX_LIBS

Fixes:
 - http://autobuild.buildroot.org/results/8bf6360a343edd10c73a756ab86cdba727f9ea43

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>