package/docker-engine: bump version to 20.10.1
Additionally, fix the ldflags specification of the package version.

The "autogen" script does not actually configure anything anymore, and
instead exports a LDFLAGS environment variable which we can't
use. Instead, specify the version information via LDFLAGS directly in
the Buildroot makefile, similar to containerd and other packages.

While at it, fix the formatting of the hash file for the LICENSE file
hash.

Signed-off-by: Christian Stewart <christian@paral.in>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>