package/php: needs pcre2
Since 2008 pcre is a hard-dependency of php:
https://github.com/php/php-src/commit/aa64c6727c2d33a647bdbd9769215044f7cf2967

Instead of optionally depending on the pcre2 package and building the
bundled pcre2 code of php in case BR2_PACKAGE_PCRE2 was not selected
we let php depend on pcre2.

While being at it rename the pcre-related configure option due to
upstream commit:
https://github.com/php/php-src/commit/c1a22f3d4e69cb87f8deeed7d03d9434e6901537

Fixes:
http://autobuild.buildroot.net/results/4a5/4a582af6b66c59a61b75a7047d8530202972ebdd/
because the pcre2 package already contains the fix for mips r6.

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>
[yann.morin.1998@free.fr: keep the JIT option]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>