package/postgresql: security bump to version 12.6
Fixes the following security issue:

- CVE-2021-3393: Partition constraint violation errors leak values of denied columns

A user having an UPDATE privilege on a partitioned table but lacking the
SELECT privilege on some column may be able to acquire denied-column values
from an error message.  This is similar to CVE-2014-8161, but the conditions
to exploit are more rare.

For more details, see the announcement:
https://www.postgresql.org/about/news/postgresql-132-126-1111-1016-9621-and-9525-released-2165/

Update the COPYRIGHT hash dur to a copyright year bump:
https://git.postgresql.org/gitweb/?p=postgresql.git;a=commitdiff;h=c09f6882d6f78bde26fcc1e1a3da11c274de596a

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>