package/qt5/qt5doc: add package with qt quick demos
Since qt 5.12, quick demos are moved in the qt5doc package.
In order to use quick demos, qt5doc has to be built.
qt5doc is a bit misleading name for the package, while it's
basic content are the quick demos. qdoc tool itself is a part
of qt5tools.

The examples need to be enabled for this package to do anything (because
we don't build the docs themselves), so select it

Signed-off-by: Goran Rađenović <gradenovic@ultratronik.de>
[Arnout:
 - fix check-package warnings;
 - update to Jan 31 2022 version of kde/5.15 branch;
 - add select of BR2_PACKAGE_QT5BASE_EXAMPLES;
 - add hash for license file]
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>