package/avrdude: Switch to upstream
In the meanwhile "linuxspi" programmer was merged upstream, therefore
it's possible to switch to latest upstream tree instead of the old fork
without losing any functionality.

The fork we were using did auto-detection of kernel headers to detect
whether spidev.h was present or not, and thus whether to enable or
disable its linuxspi 'driver'. As spidev.h has been present since
linux-2.6.22, we can quite easily conclude that spidev support was
always enabled in the fork.

But upstream went a slightly different route, and added a configure
option (and they do no validity check).

So, to keep backward behaviour, we unconditionally enable it  now that
we switched back to use upstream.

Signed-off-by: Alexander Sverdlin <alexander.sverdlin@gmail.com>
[yann.morin.1998@free.fr:
  - clarify why we forcibly use --enable-linuxspi
  - fix the hash to adapt to the new svn tarball format (c043ecb20ce6)
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>