package/mender: do not automatically select uboot tools if uboot is selected
Because uboot requires a set of unique patches for each board, the
upstream package developers are phasing out supporting uboot wherever
possible.  Instead, they recommend using Grub2 as a secondary
bootloader and using the mender-grubenv package.

Because the mender-grubenv file provides it's own fw_printenv script,
it is not possible to know if U-Boot's fw_printenv or mender-grubenv's
fw_printenv should be used.

As such, remove selecting uboot tools when uboot is selected, and
instead add a more comprehensive note in the help section about what
Mender requires for uboot and grub2-based systems, with a link to the
meta-mender github project for the base uboot patches, and a link to
the official documentation on manual uboot integration.

Signed-off-by: Adam Duskett <aduskett@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>