package/tvheadend: add option to enable IPTV support
Some IPTV streams will need ffmpef, the command line tool), while some
won't, so we just suggest that to the user in the help text.

There were two alternatives, but neither were very convincing:
  - always enforce that ffmpeg is enabled
  - only enforce ffmpeg to be enabled when the package is already
    enabled

In either case, that may cary the ffmpeg tool when it really is not
needed. So leave it to the user to decide whether they want it or not.

tvheadend now has a bunch of options, so make it a sub-menu.

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>
[yann.morin.1998@free.fr:
  - do not forcibly enable ffmpeg-the-commandline-tool
  - one option per-commit
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>