package/timescaledb: bump version to 2.0.2
Bugfix release with important bugfixes:
https://github.com/timescale/timescaledb/releases/tag/2.0.2

This maintenance release contains bugfixes since the 2.0.1 release. We
deem it high priority for upgrading.

The bug fixes in this release address issues with joins, the status of
background jobs, and disabling compression. It also includes
enhancements to continuous aggregates, including improved validation
of policies and optimizations for faster refreshes when there are a
lot of invalidations.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>